Update MultiscaleImage to support update somacore spec
This required refactoring `BaseCollection` so `MultiscaleImage` and
`BaseCollection` both inherit from a new `SOMAGroup`, but
`MultiscaleImage` does not inherit the collection-specific methods.
