+++ apis/python/src/tiledbsoma/io/ingest.py
@@ -46,7 +46,6 @@ from .. import (
     DenseNDArray,
     Experiment,
     Measurement,
-    MultiscaleImage,
     PointCloud,
     Scene,
     SparseNDArray,
@@ -1004,17 +1003,6 @@ def _create_or_open_collection(
     context: Optional["SOMATileDBContext"],
     additional_metadata: "AdditionalMetadata" = None,
 ) -> Scene: ...
-
-
-@overload
-def _create_or_open_collection(
-    cls: Type[MultiscaleImage],
-    uri: str,
-    *,
-    ingestion_params: IngestionParams,
-    context: Optional["SOMATileDBContext"],
-    additional_metadata: "AdditionalMetadata" = None,
-) -> MultiscaleImage: ...
 
 
 @no_type_check
